Renamed CharacterControl to PlayerControl to avoid confusion with the Unity component; Added namespace to all scripts
