+++ Assets/Scripts/Build.cs
@@ -2,15 +2,16 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Build : MonoBehaviour {
+namespace TowerDefense {
+    public class Build : MonoBehaviour {
 
-	public GameObject BuildStructure(Vector3 spawnLocation, Quaternion spawnRotation, GameObject spawnPrefab) {
+		public GameObject BuildStructure(Vector3 spawnLocation, Quaternion spawnRotation, GameObject spawnPrefab) {
 
-		float spawnY = spawnPrefab.GetComponent<BoxCollider>().size.y;
-		Vector3 spawnOffset = spawnLocation - new Vector3(0, spawnY / 2, 0);
-		spawnLocation = spawnLocation + new Vector3(0, spawnY / 2, 0);
+			float spawnY = spawnPrefab.GetComponent<BoxCollider>().size.y;
+			Vector3 spawnOffset = spawnLocation - new Vector3(0, spawnY / 2, 0);
+			spawnLocation = spawnLocation + new Vector3(0, spawnY / 2, 0);
 
-		return (GameObject)Instantiate(spawnPrefab, spawnOffset, spawnRotation);
+			return (GameObject)Instantiate(spawnPrefab, spawnOffset, spawnRotation);
+		}
 	}
-
 }
