+++ Assets/Scripts/Entity.cs
@@ -2,15 +2,17 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Entity : MonoBehaviour {
+namespace TowerDefense {
+	public class Entity : MonoBehaviour {
 
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
+		// Use this for initialization
+		void Start () {
+
+		}
+
+		// Update is called once per frame
+		void Update () {
+
+		}
 	}
 }
