+++ Assets/Scripts/Health.cs
@@ -1,71 +1,73 @@
 using System.Collections;
 using UnityEngine;
 
-public class Health : MonoBehaviour {
+namespace TowerDefense {
+	public class Health : MonoBehaviour {
 
-	public float maxHealth;
+		public float maxHealth;
 
-	private float flashTime = 0.05f;
-	private Color hitColor = new Color(1f, 0f, 0f, 1f);
-	private Color normalColor;
-	private MeshRenderer rendy;
-	private MeshRenderer[] rendsArray;
+		private float flashTime = 0.05f;
+		private Color hitColor = new Color(1f, 0f, 0f, 1f);
+		private Color normalColor;
+		private MeshRenderer rendy;
+		private MeshRenderer[] rendsArray;
 
-	[SerializeField]
-	private float currentHealth;
+		[SerializeField]
+		private float currentHealth;
 
-	public float CurrentHealth {
-		get { return currentHealth; }
-		set {
-			currentHealth = value;
+		public float CurrentHealth {
+			get { return currentHealth; }
+			set {
+				currentHealth = value;
+			}
 		}
-	}
 
-	[SerializeField]
-	private bool allegiance; // True = Friendly, False = Enemy
+		[SerializeField]
+		private bool allegiance; // True = Friendly, False = Enemy
 
-	public bool Allegiance {
-		get { return allegiance; }
-		set {
-			allegiance = value;
+		public bool Allegiance {
+			get { return allegiance; }
+			set {
+				allegiance = value;
+			}
 		}
-	}
 
-	[SerializeField]
-	private float targetBias;
+		[SerializeField]
+		private float targetBias;
 
-	public float TargetBias {
-		get { return targetBias; }
-	}
-
-	void Awake() {
-		SetInitialReferences();
-	}
-
-	void SetInitialReferences() {
-		rendy = gameObject.GetComponentsInChildren<MeshRenderer>()[0];
-		normalColor = rendy.material.color;
-
-		currentHealth = maxHealth;
-	}
-
-	public void TakeDamage(float damage) {
-		currentHealth -= damage;
-		if(currentHealth <= 0) {
-			Die();
-		} else {
-			StartCoroutine(Flash());
+		public float TargetBias {
+			get { return targetBias; }
 		}
-	}
 
-	public void Die() {
-		// Separate out and check whether GO is player
-		Destroy(gameObject);
-	}
+		void Awake() {
+			SetInitialReferences();
+		}
 
-	IEnumerator Flash() {
-		rendy.material.color = hitColor;
-		yield return new WaitForSeconds(flashTime);
-		rendy.material.color = normalColor;
+		void SetInitialReferences() {
+			rendy = gameObject.GetComponentsInChildren<MeshRenderer>()[0];
+			normalColor = rendy.material.color;
+
+			currentHealth = maxHealth;
+		}
+
+		public void TakeDamage(float damage) {
+			currentHealth -= damage;
+			if(currentHealth <= 0) {
+				Die();
+			} else {
+				StartCoroutine(Flash());
+			}
+		}
+
+		public void Die() {
+			// Separate out and check whether GO is player
+			Destroy(gameObject);
+		}
+
+		IEnumerator Flash() {
+			rendy.material.color = hitColor;
+			yield return new WaitForSeconds(flashTime);
+			rendy.material.color = normalColor;
+		}
 	}
 }
